requirements.txt
reamber==0.1.6
scikit-learn
junit_xml
pandas
pytorch-lightning
